# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CIOS Forecasting Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all forecast cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new forecast case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{caseId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a specific case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    put:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /cases/{caseId}/forecast:
    get:
      operationId: runForecast
      tags: [forecasts]
      summary: Run the CIOS forecast engine for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Forecast result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastResult"
...
  /actors/{actorId}:
    put:
      operationId: updateActor
      tags: [actors]
      summary: Update actor weights
      parameters:
        - name: actorId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActorRequest"
      responses:
        "200":
          description: Updated actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
  /case-library:
    get:
      operationId: listCaseLibrary
      tags: [analogs]
      summary: List historical analog cases
      responses:
        "200":
          description: Case library
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalogCase"
    post:
      operationId: createAnalogCase
      tags: [analogs]
      summary: Add an analog case to the library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalogCaseRequest"
      responses:
        "201":
          description: Created analog case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalogCase"
  /case-library/{analogId}:
    put:
      operationId: updateAnalogCase
      tags: [analogs]
      summary: Update an analog case
      parameters:
        - name: analogId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalogCaseRequest"
      responses:
        "200":
          description: Updated analog case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalogCase"
    delete:
      operationId: deleteAnalogCase
      tags: [analogs]
      summary: Delete an analog case
      parameters:
        - name: analogId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /cases/{caseId}/analogs:
    get:
      operationId: getAnalogs
      tags: [analogs]
      summary: Retrieve top analog cases for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ranked analog matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalogMatch"
...
  /calibration/{forecastId}/outcome:
    post:
      operationId: recordOutcome
      tags: [calibration]
      summary: Record actual outcome for a forecast
      parameters:
        - name: forecastId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordOutcomeRequest"
      responses:
        "200":
          description: Updated calibration entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalibrationEntry"
...
  /guidance:
    get:
      operationId: listGuidance
      tags: [guidance]
      summary: List strategic guidance items
      parameters:
        - name: caseId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Guidance items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuidanceItem"
    post:
      operationId: createGuidance
      tags: [guidance]
      summary: Create a guidance item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuidanceRequest"
      responses:
        "201":
          description: Created guidance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuidanceItem"
...
components:
  schemas:
...
    Case:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        strategicQuestion:
          type: string
        outcomeDefinition:
          type: string
        timeHorizon:
          type: string
        priorProbability:
          type: number
        currentProbability:
          type: number
        confidenceLevel:
          type: string
        primaryBrand:
          type: string
        primarySpecialtyProfile:
          type: string
        payerEnvironment:
          type: string
        guidelineLeverage:
          type: string
        competitorProfile:
          type: string
        lastUpdate:
          type: string
        signalCount:
          type: integer
        topSupportiveActor:
          type: string
          nullable: true
        topConstrainingActor:
          type: string
          nullable: true
        miosRoutingCheck:
          type: string
          nullable: true
        ohosRoutingCheck:
          type: string
          nullable: true
      required:
        - id
        - caseId
        - strategicQuestion
        - priorProbability
    CreateCaseRequest:
      type: object
      properties:
        caseId:
          type: string
        strategicQuestion:
          type: string
        outcomeDefinition:
          type: string
        timeHorizon:
          type: string
        priorProbability:
          type: number
        primaryBrand:
          type: string
        primarySpecialtyProfile:
          type: string
        payerEnvironment:
          type: string
        guidelineLeverage:
          type: string
        competitorProfile:
          type: string
      required:
        - strategicQuestion
        - priorProbability
        - primaryBrand
...
    Actor:
      type: object
      properties:
        id:
          type: string
        actorName:
          type: string
        influenceWeight:
          type: number
        positiveResponseFactor:
          type: number
        negativeResponseFactor:
          type: number
        outcomeOrientation:
          type: number
        roleInSystem:
          type: string
          nullable: true
        baseInfluenceWeight:
          type: number
        canonicalActor:
          type: string
          nullable: true
        specialtyProfile:
          type: string
          nullable: true
      required:
        - id
        - actorName
        - influenceWeight
        - positiveResponseFactor
        - negativeResponseFactor
        - outcomeOrientation
    UpdateActorRequest:
      type: object
      properties:
        influenceWeight:
          type: number
        positiveResponseFactor:
          type: number
        negativeResponseFactor:
          type: number
        roleInSystem:
          type: string
      required:
        - influenceWeight
        - positiveResponseFactor
        - negativeResponseFactor
    ForecastResult:
      type: object
      properties:
        caseId:
          type: string
        priorProbability:
          type: number
        priorOdds:
          type: number
        signalLrProduct:
          type: number
        actorAdjustmentFactor:
          type: number
        posteriorOdds:
          type: number
        currentProbability:
          type: number
        confidenceLevel:
          type: string
        netActorTranslation:
          type: number
        topSupportiveActor:
          type: string
          nullable: true
        topConstrainingActor:
          type: string
          nullable: true
        actorAggregation:
          type: array
          items:
            $ref: "#/components/schemas/ActorAggregation"
        signalDetails:
          type: array
          items:
            $ref: "#/components/schemas/SignalForecastDetail"
        interpretation:
          $ref: "#/components/schemas/ForecastInterpretation"
        forecastId:
          type: string
        savedAt:
          type: string
      required:
        - caseId
        - priorProbability
        - currentProbability
        - confidenceLevel
    ActorAggregation:
      type: object
      properties:
        actor:
          type: string
        netActorEffect:
          type: number
        interpretation:
          type: string
      required:
        - actor
        - netActorEffect
        - interpretation
    SignalForecastDetail:
      type: object
      properties:
        signalId:
          type: string
        description:
          type: string
        direction:
          type: string
        likelihoodRatio:
          type: number
        weightedActorReaction:
          type: number
        actorReactions:
          type: object
          additionalProperties:
            type: number
      required:
        - signalId
        - description
        - direction
        - likelihoodRatio
    ForecastInterpretation:
      type: object
      properties:
        primaryStatement:
          type: string
        topSupportiveActor:
          type: string
          nullable: true
        topConstrainingActor:
          type: string
          nullable: true
        highestImpactSignal:
          type: string
          nullable: true
        recommendedAction:
          type: string
          nullable: true
        miosRoutingCheck:
          type: string
          nullable: true
        ohosRoutingCheck:
          type: string
          nullable: true
        behavioralSummary:
          type: string
          nullable: true
      required:
        - primaryStatement
    AnalogCase:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        therapyArea:
          type: string
        productType:
          type: string
        specialty:
          type: string
        evidenceType:
          type: string
          nullable: true
        lifecycleStage:
          type: string
          nullable: true
        actorMix:
          type: string
          nullable: true
        marketAccessConditions:
          type: string
          nullable: true
        outcomePattern:
          type: string
          nullable: true
        adoptionTrajectory:
          type: string
          nullable: true
        keyInflectionSignals:
          type: string
          nullable: true
        finalObservedOutcome:
          type: string
          nullable: true
        finalProbability:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - caseId
        - therapyArea
        - productType
        - specialty
    CreateAnalogCaseRequest:
      type: object
      properties:
        caseId:
          type: string
        therapyArea:
          type: string
        productType:
          type: string
        specialty:
          type: string
        evidenceType:
          type: string
        lifecycleStage:
          type: string
        actorMix:
          type: string
        marketAccessConditions:
          type: string
        outcomePattern:
          type: string
        adoptionTrajectory:
          type: string
        keyInflectionSignals:
          type: string
        finalObservedOutcome:
          type: string
        finalProbability:
          type: number
        notes:
          type: string
      required:
        - caseId
        - therapyArea
        - productType
        - specialty
    AnalogMatch:
      type: object
      properties:
        analogCase:
          $ref: "#/components/schemas/AnalogCase"
        similarityScore:
          type: number
        similarityReasoning:
          type: string
        matchedDimensions:
          type: array
          items:
            type: string
      required:
        - analogCase
        - similarityScore
        - similarityReasoning
    CalibrationEntry:
      type: object
      properties:
        id:
          type: string
        forecastId:
          type: string
        caseId:
          type: string
        predictionDate:
          type: string
        predictedProbability:
          type: number
        observedOutcome:
          type: number
          nullable: true
        brierComponent:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        userFeedback:
          type: string
          nullable: true
        reviewerComments:
          type: string
          nullable: true
        forecastError:
          type: number
          nullable: true
      required:
        - id
        - forecastId
        - caseId
        - predictionDate
        - predictedProbability
    RecordOutcomeRequest:
      type: object
      properties:
        observedOutcome:
          type: number
        notes:
          type: string
        userFeedback:
          type: string
        reviewerComments:
          type: string
      required:
        - observedOutcome
...
    GuidanceItem:
      type: object
      properties:
        id:
          type: string
        guidanceId:
          type: string
        caseId:
          type: string
        keyRiskDriver:
          type: string
        recommendedAction:
          type: string
        targetAudience:
          type: string
          nullable: true
        priorityLevel:
          type: string
        linkedSignalId:
          type: string
          nullable: true
      required:
        - id
        - guidanceId
        - caseId
        - keyRiskDriver
        - recommendedAction
        - priorityLevel
    CreateGuidanceRequest:
      type: object
      properties:
        caseId:
          type: string
        keyRiskDriver:
          type: string
        recommendedAction:
          type: string
        targetAudience:
          type: string
        priorityLevel:
          type: string
        linkedSignalId:
          type: string
      required:
        - caseId
        - keyRiskDriver
        - recommendedAction
        - priorityLevel
...