# lib/api-zod/src/generated/types/calibrationEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CIOS Forecasting Engine API
 * OpenAPI spec version: 1.0.0
 */

export interface CalibrationEntry {
  id: string;
  forecastId: string;
  caseId: string;
  predictionDate: string;
  predictedProbability: number;
  observedOutcome?: number | null;
  brierComponent?: number | null;
  notes?: string | null;
  userFeedback?: string | null;
  reviewerComments?: string | null;
  forecastError?: number | null;
}