# lib/api-zod/src/generated/types/updateActorRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CIOS Forecasting Engine API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateActorRequest {
  influenceWeight: number;
  positiveResponseFactor: number;
  negativeResponseFactor: number;
  roleInSystem?: string;
}